Static Analysis fix for qcnode

+++ source/Library/VideoCodec/VidcDrvClient.cpp
@@ -278,7 +278,7 @@ QCStatus_e VidcDrvClient::LoadResources()
     QC_DEBUG( "Loading vidc resources" );
     InitCmdCompleted();
     rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_LOAD_RESOURCES, nullptr, 0, nullptr, 0 );
-    if ( VIDC_ERR_NONE != rc )
+    if ( static_cast<int32_t>(VIDC_ERR_NONE) != rc )
     {
         QC_ERROR( "Loading vidc resources failed! rc=0x%x, %s", rc,
                   VidcErrToStr( vidc_status_type( rc ) ) );
@@ -307,7 +307,7 @@ QCStatus_e VidcDrvClient::ReleaseResources()
     QC_DEBUG( "release vidc resources" );
     InitCmdCompleted();
     rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_RELEASE_RESOURCES, nullptr, 0, nullptr, 0 );
-    if ( VIDC_ERR_NONE != rc )
+    if ( static_cast<int32_t>(VIDC_ERR_NONE) != rc )
     {
         QC_ERROR( "release vidc resources failed! rc=0x%x, %s", rc,
                   VidcErrToStr( vidc_status_type( rc ) ) );
@@ -478,7 +478,7 @@ QCStatus_e VidcDrvClient::StartDriver( VideoCodec_StartType_e type )
         rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_START, nullptr, 0, nullptr, 0 );
     }
 
-    if ( VIDC_ERR_NONE != rc )
+    if (static_cast<int32_t>(VIDC_ERR_NONE) != rc )
     {
         QC_ERROR( "Starting vidc failed! rc=0x%x %s", rc, VidcErrToStr( vidc_status_type( rc ) ) );
         ret = QC_STATUS_FAIL;
@@ -535,7 +535,7 @@ QCStatus_e VidcDrvClient::SetBuffer( VideoCodec_BufType_e bufferType,
 #endif
         buf_info.buf_type = vidcBufType;
         buf_info.contiguous = true;
-        buf_info.buf_size = buf.size;
+        buf_info.buf_size = static_cast<uint32>(buf.size);
         // register it before lookup in future for local allocation
         QC_DEBUG( "set-buffer to driver [%" PRId32 "]: buf_addr = 0x%x, buf_handle = 0x%x, "
                  "type:%d, buf_size:%d",
@@ -544,7 +544,7 @@ QCStatus_e VidcDrvClient::SetBuffer( VideoCodec_BufType_e bufferType,
         rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_SET_BUFFER, (uint8_t*) (&buf_info),
                            sizeof(vidc_buffer_info_type),
                            nullptr, 0 );
-        if (VIDC_ERR_NONE != rc)
+        if (static_cast<int32_t>(VIDC_ERR_NONE) != rc)
         {
             QC_ERROR( "set-buffer VIDC_IOCTL_SET_BUFFER failed. Index=%" PRId32 " rc=0x%x", i,
                             rc );
@@ -605,12 +605,12 @@ QCStatus_e VidcDrvClient::FreeBuffers( VideoCodec_BufType_e bufferType,
 #endif
         buf_info.buf_type = vidcBufType;
         buf_info.contiguous = true;
-        buf_info.buf_size = buf.size;
+        buf_info.buf_size = static_cast<uint32>(buf.size);
 
         QC_DEBUG( "FREE_BUFFER, i:%d, size:%d", i, buf_info.buf_size );
         rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_FREE_BUFFER, (uint8_t *) ( &buf_info ),
                            sizeof( vidc_buffer_info_type ), nullptr, 0 );
-        if ( VIDC_ERR_NONE != rc )
+        if ( static_cast<int32_t>(VIDC_ERR_NONE) != rc )
         {
             ret = QC_STATUS_FAIL;
             QC_ERROR( "VIDC_IOCTL_FREE_BUFFER index=%" PRIu32 ", type:%" PRIu32
@@ -708,7 +708,7 @@ QCStatus_e VidcDrvClient::EmptyBuffer( VideoFrameDescriptor &frameDesc )
     {
         rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_EMPTY_INPUT_BUFFER, (uint8_t *) ( &frameData ),
                            sizeof( frameData ), nullptr, 0 );
-        if ( VIDC_ERR_NONE == rc )
+        if ( static_cast<int32_t>(VIDC_ERR_NONE) == rc )
         {
             QC_DEBUG( "SubmitInputFrame VIDC_IOCTL_EMPTY_INPUT_BUFFER succeeded. frameTs: %" PRIu64
                       " inputMapSsize: %zu", timestampUs, m_inputMap.size() );
@@ -808,7 +808,7 @@ QCStatus_e VidcDrvClient::FillBuffer( VideoFrameDescriptor &frameDesc )
 
         rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_FILL_OUTPUT_BUFFER, (uint8_t *) ( &frameData ),
                            sizeof( frameData ), nullptr, 0 );
-        if ( VIDC_ERR_NONE != rc )
+        if ( static_cast<int32_t>(VIDC_ERR_NONE) != rc )
         {
             QC_ERROR( "FillBuffer 0x%x failed rc 0x%x", handle, rc );
             m_outputMap[handle].bUsedFlag = false;
@@ -837,7 +837,7 @@ QCStatus_e VidcDrvClient::StopDecoder()
     InitCmdCompleted();
     m_bCmdDrainReceived = false;
     rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_DRAIN, nullptr, 0, nullptr, 0 );
-    if ( VIDC_ERR_NONE != rc )
+    if ( static_cast<int32_t>(VIDC_ERR_NONE) != rc )
     {
         QC_ERROR( "Stop vidc failed! rc=0x%x", rc );
         ret = QC_STATUS_FAIL;
@@ -859,7 +859,7 @@ QCStatus_e VidcDrvClient::StopDecoder()
         stop_mode = VIDC_STOP_INPUT;
         rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_STOP, (uint8 *) &stop_mode,
                            sizeof( vidc_stop_mode_type ), nullptr, 0 );
-        if ( VIDC_ERR_NONE != rc )
+        if ( static_cast<int32_t>(VIDC_ERR_NONE) != rc )
         {
             QC_ERROR( "vidc stop input failed! rc=0x%x", rc );
             ret = QC_STATUS_FAIL;
@@ -881,7 +881,7 @@ QCStatus_e VidcDrvClient::StopDecoder()
         stop_mode = VIDC_STOP_OUTPUT;
         rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_STOP, (uint8 *) &stop_mode,
                            sizeof( vidc_stop_mode_type ), nullptr, 0 );
-        if ( VIDC_ERR_NONE != rc )
+        if ( static_cast<int32_t>(VIDC_ERR_NONE) != rc )
         {
             QC_ERROR( "vidc stop output failed! rc=0x%x", rc );
             ret = QC_STATUS_FAIL;
@@ -906,7 +906,7 @@ QCStatus_e VidcDrvClient::StopEncoder()
 
     InitCmdCompleted();
     int32_t rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_STOP, nullptr, 0, nullptr, 0 );
-    if ( VIDC_ERR_NONE != rc )
+    if ( static_cast<int32_t>(VIDC_ERR_NONE) != rc )
     {
         QC_ERROR( "Stop vidc failed! rc=0x%x", rc );
         ret = QC_STATUS_FAIL;
@@ -1040,7 +1040,7 @@ int VidcDrvClient::DeviceCbHandler( uint8_t *pMsg, uint32_t length )
                 }
             }
 
-            if ( ( pFrameData->flags & VIDC_FRAME_FLAG_EOS ) > 0 )
+            if ( ( pFrameData->flags & static_cast<unsigned int>(VIDC_FRAME_FLAG_EOS) ) > 0 )
             {
                 QC_WARN( "detected VIDC_FRAME_FLAG_EOS -- not possible for camera streaming" );
             }
@@ -1163,7 +1163,7 @@ QCStatus_e VidcDrvClient::GetDrvProperty( uint32_t id, uint32_t nPktSize, const 
     QCStatus_e ret = QC_STATUS_OK;
     uint8_t dev_cmd_buffer[VIDEO_MAX_DEV_CMD_BUFFER_SIZE] = { 0 };
     vidc_drv_property_type *pProp = (vidc_drv_property_type *) dev_cmd_buffer;
-    uint32_t nMsgSize = sizeof( vidc_property_hdr_type ) + nPktSize;
+    uint32_t nMsgSize = static_cast<uint32_t>(sizeof( vidc_property_hdr_type ) + nPktSize);
     vidc_property_id_type propId = (vidc_property_id_type) id;
 
     // pProp->payload buffer is more than 1 byte as the struct defined;
@@ -1180,9 +1180,9 @@ QCStatus_e VidcDrvClient::GetDrvProperty( uint32_t id, uint32_t nPktSize, const 
         pProp->prop_hdr.size = nPktSize;
         pProp->prop_hdr.prop_id = propId;
 
-        rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_GET_PROPERTY, dev_cmd_buffer, (int32_t) nMsgSize,
+        rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_GET_PROPERTY, dev_cmd_buffer, static_cast<uint32_t>(nMsgSize),
                            (uint8_t*) &pkt, nPktSize );
-        if ( VIDC_ERR_NONE != rc )
+        if ( static_cast<int32_t>(VIDC_ERR_NONE) != rc )
         {
             printf( "GetDrvProperty propId=0x%x failed! rc=0x%x\n", propId, rc );
             ret = QC_STATUS_FAIL;
@@ -1198,7 +1198,7 @@ QCStatus_e VidcDrvClient::SetDrvProperty( uint32_t id, uint32_t nPktSize, const 
     QCStatus_e ret = QC_STATUS_OK;
     uint8_t dev_cmd_buffer[VIDEO_MAX_DEV_CMD_BUFFER_SIZE] = { 0 };
     vidc_drv_property_type *pProp = (vidc_drv_property_type *) dev_cmd_buffer;
-    uint32_t nMsgSize = sizeof( vidc_property_hdr_type ) + nPktSize;
+    uint32_t nMsgSize = static_cast<uint32_t>(sizeof( vidc_property_hdr_type ) + nPktSize);
     vidc_property_id_type propId = (vidc_property_id_type) id;
 
     // pProp->payload buffer is more than 1 byte as the struct defined;
@@ -1216,9 +1216,9 @@ QCStatus_e VidcDrvClient::SetDrvProperty( uint32_t id, uint32_t nPktSize, const 
         pProp->prop_hdr.size = nPktSize;
         pProp->prop_hdr.prop_id = propId;
 
-        rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_SET_PROPERTY, dev_cmd_buffer, (int32_t) nMsgSize,
+        rc = device_ioctl( m_pIoHandle, VIDC_IOCTL_SET_PROPERTY, dev_cmd_buffer, static_cast<uint32_t>(nMsgSize),
                            nullptr, 0 );
-        if ( VIDC_ERR_NONE != rc )
+        if ( static_cast<int32_t>(VIDC_ERR_NONE) != rc )
         {
             QC_ERROR( "SetDrvProperty propId=0x%x failed! rc=0x%x, %s", propId, rc,
                       VidcErrToStr( vidc_status_type( rc ) ) );
